+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2023-02-06
+
+### Added
+
+- Registry support. The default registry is <https://registry.litebds.com>.
+
+### Fixed
+
+- Failing to uninstall teeth with uppercase letters in provided tooth path.
+
 ## [0.7.1] - 2023-02-05
 
 ### Fixed
@@ -121,7 +131,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic functions: cache, install, list, show, tooth init, and uninstall.
 
-[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.7.1...HEAD
+[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/LiteLDev/Lip/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/LiteLDev/Lip/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/LiteLDev/Lip/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/LiteLDev/Lip/compare/v0.5.1...v0.6.0